requirements.txt
asgiref==3.8.1
coverage==7.4.4
Django==5.0.4
djangorestframework==3.15.1
sqlparse==0.4.4
tzdata==2024.1
